feat: adding router to get all services
